requirements.txt
einops==0.2.0
paddlepaddle
paddleocr
pyclipper
lmdb
numpy
pandas
opencv-python
pillow
matplotlib
torch
underthesea>=1.3.1
